requirements.txt
streamlit
opencv-python-headless
numpy
pandas
Pillow
svglib
reportlab
plotly